+++ docs/topics/release-notes.md
@@ -28,6 +28,7 @@ You can determine your currently installed version using `pip freeze`:
 
 ### Master
 
+* Added a `post_save()` hook to the generic views.
 * Bugfix: Fix styling on browsable API login.
 * Bugfix: Fix issue with deserializing empty to-many relations.
 * Bugfix: Ensure model field validation is still applied for ModelSerializer subclasses with an custom `.restore_object()` method.